producer/requirements.txt
# producer/requirements.txt

# Kafka client
kafka-python

# AWS SDK for Python (for S3) - Constrained for Hetzner compatibility
# Use versions < 1.35 based on previous troubleshooting, adjust if needed
boto3<1.35
botocore<1.35

# Image processing library (for process_new_images.py helper)
opencv-python

# Filesystem monitoring (for continuous producer mode)
watchdog

# Loading .env files for configuration
python-dotenv

# For video download/extraction helpers (yt-dlp)
yt-dlp